Nuke and pave the Cards module.
Instead of having a Card limited to a standard rank-and-suit card, we
now have Card<T> where T: CardFaces. This allows us to work with cards
with any type of content front-and-back.

Standard playing cards now defined in their own module
(std_playing_cards.rs) and can still be used in much the same way as in
previous version, creating a deck of Card<StandardCard>. Jokers are now
available and treated properly as wildcards in...

Hand<StandardCard> implementation adds analytic methods to check for
flushes/n-of-a-kind/straights and can provide handy rank and suit
frequency maps to aid in further analytics.